requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
radon>=6.0.1
bandit>=1.7.9
jinja2>=3.1.4
